requirements.txt
speechrecognition
pyaudio
pyttsx3
python-dotenv
loguru
ollama
langchain
langchain-community
